move asr, download_limit and settings to a lifespan function
